Reimplement everyNth as sugar on top of filterIndexed, rename to takeEveryNth and dropEveryNth
