+++ docs/gatherers4j/content/gatherers/filtering-and-selection/dropLast.md
@@ -1,0 +1,31 @@
+---
+title: "dropLast()"
+linkTitle: "dropLast()"
+show_in_table: true
+description: Keep all elements except the last `count` elements of the stream.
+---
+
+
+### Implementation Notes
+
+Holds at most `count` number of elements in memory before emitting them as it becomes clear that they are not in the trailing `count` elements in the stream.
+
+**Signature**
+
+`dropLast(int count)`
+
+* `count` - The positive number of elements to remove from the end of the input stream
+
+### Examples
+
+#### Drop the last 2 elements from the input stream
+
+```java
+Stream
+    .of("A", "B", "C", "D", "E")
+    .gather(Gatherers4j.dropLast(2))
+    .toList();
+
+// ["A", "B", "C"]
+```
+
